Transformed the script into a blender addon.
