--- projects/MarlQ/src/blender.bat
@@ -1,3 +1,0 @@
-@echo off
-"G:\Geometry Processing\Blender 2.93\blender.exe" --factory-startup  -b -P .\main.py
-pause
